instrumentation-genai/opentelemetry-instrumentation-weaviate/test-requirements-v3.txt
weaviate-client>=3.0.0,<4.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
pytest-vcr==1.0.2
-e opentelemetry-instrumentation
-e instrumentation-genai/opentelemetry-instrumentation-weaviate